Making it work...
We want to solve a couple of problems:

1) You don't want to create a separate project with a JAR to be shared between teams

the problem that you have is that the Gradle plugin classpath is different to the one that you have in your sources. Now, you would like to reuse your sources in the Gradle plugin classpath (when Spring Cloud Contract plugin is executed). In order to achiee that you need to pass the sources that you'd like to use in your Gradle plugin classpath to the `buildSrc` folder. It's a special folder that gets added to the classpath. That way the plugin will start understanding the additional files.

2) ... but you would like to have code completion in your test sources ....

 Your contracts are  on your test classpath. The cool thing about SC-Contract is that you have code completion. To have it you need all the sources present. In your case since the sources lay in the `buildSrc` we need to manually tell Gradle to add the compiled sources from `buildSrc` to your test classpath

 3) ... and unfortunately only you will be able to read those contracts - for all other teams everything will blow up

 The contracts are re-read by the clients who want to profit from messaging for example. Now imagine putting a reference to a class that is only available in your `buildSrc`. Yeah - it will blow up. That's why you can use this approach only if you want to play around and don't want to share your contracts anywhere.